requirements/common.txt
Django
argon2-cffi
djangorestframework
django-configurations
django-cors-headers
django-storages
psycopg2
